Add a DynamicView to setup Screens from a (JSON) Configuration (#71)
A DynamicView has been implemented to be able to setup screen from a (JSON)
configuration files.

The idea is to provide several widgets that the user can choose from in
a (JSON) config. A DynamicViewModel needs to be implemented that's
responsible for reading in the (JSON) configuration which the ViewModel
then provides to a DynamicView. The actual Widgets _(composables)_ from
the configuration are then displayed by the DynamicView.
Each widget triggers an action that can be observed by the ViewModel
instance.