ui-toolkit: Add WLANManager for the WifiWidget (#82)
## What's new?

Added a wlan manger to scan for a specific wlan or/and connect to it via
a given ssid. Since restrictions are given for wlan management, a
usecase for each case (depending on the api level and the given
restrictions) is given. The manager picks the right case for the given
api level. This is needed to show and handle button actions of the
connect wlan widget.

Co-authored-by: Christian Maier <christian.maier@snabble.io>